+++ contracts/src/interfaces/IRMRKMultiAsset.sol
@@ -136,7 +136,7 @@ interface IRMRKMultiAsset is IERC165, IRMRKMultiAssetEventsAndStruct {
 
     /**
      * @notice Sets a new priority array for a given token.
-     * @dev The priority array is a non-sequential list of `uint16`s, where the lowest value is considered highest
+     * @dev The priority array is a non-sequential list of `uint64`s, where the lowest value is considered highest
      *  priority.
      * @dev Value `0` of a priority is a special case equivalent to unitialized.
      * @dev Requirements:
@@ -149,7 +149,7 @@ interface IRMRKMultiAsset is IERC165, IRMRKMultiAssetEventsAndStruct {
      * @param priorities An array of priorities of active assets. The succesion of items in the priorities array
      *  matches that of the succesion of items in the active array
      */
-    function setPriority(uint256 tokenId, uint16[] calldata priorities)
+    function setPriority(uint256 tokenId, uint64[] calldata priorities)
         external;
 
     /**
@@ -179,15 +179,15 @@ interface IRMRKMultiAsset is IERC165, IRMRKMultiAssetEventsAndStruct {
 
     /**
      * @notice Used to retrieve the priorities of the active resoources of a given token.
-     * @dev Asset priorities are a non-sequential array of uint16 values with an array size equal to active asset
+     * @dev Asset priorities are a non-sequential array of uint64 values with an array size equal to active asset
      *  priorites.
      * @param tokenId ID of the token for which to retrieve the priorities of the active assets
-     * @return uint16[] An array of priorities of the active assets of the given token
+     * @return uint64[] An array of priorities of the active assets of the given token
      */
     function getActiveAssetPriorities(uint256 tokenId)
         external
         view
-        returns (uint16[] memory);
+        returns (uint64[] memory);
 
     /**
      * @notice Used to retrieve the asset that will be replaced if a given asset from the token's pending array
